Moved Spark Streaming CRD specific code into sparkStreaming.py
