+++ api/src/main/resources/application_creator.py
@@ -80,8 +80,8 @@ class ApplicationCreator(object):
                                                    property_overrides.get(component_type))
                 create_metadata[component_type] = result
         finally:
-        # clean up staged package data
-        # shutil.rmtree(stage_path)
+            #clean up staged package data
+            #shutil.rmtree(stage_path)
             pass
 
         return create_metadata
