+++ api/src/main/resources/plugins/sparkStreaming.py
@@ -25,6 +25,8 @@ either express or implied.
 import json
 import os
 import logging
+from kubernetes.client.rest import ApiException
+from kubernetes import config, client
 from shutil import copy
 import yaml
 import deployer_utils
@@ -70,6 +72,53 @@ class SparkStreamingCreator(Common):
                 rpl_data['spec']['executor']['memory'] = executor_memory
                 json.dump(rpl_data, json_out, ensure_ascii=False, indent="\t")
         return json_file
+
+    def create_custom_resource_object(self, spark_app_json_path):
+
+        config.load_incluster_config()
+        configuration = client.Configuration()
+        api_instance = client.CustomObjectsApi(client.ApiClient(configuration))
+        group = 'sparkoperator.k8s.io'  # str | The custom resource's group name
+        version = 'v1beta2'  # str | The custom resource's version
+        namespace = 'pnda'  # str | The custom resource's namespace
+        plural = 'sparkapplications'  # str | The custom resource's plural name. For TPRs this would be lowercase plural kind.
+        pretty = 'true'
+        with open(spark_app_json_path, 'r') as jfile:
+
+            data = json.load(jfile)
+
+        try:
+            api_response = api_instance.create_namespaced_custom_object(group, version, namespace, plural, body=data, pretty=pretty)
+            logging.info("CRD create response %s" % api_response)
+        except ApiException as e:
+            logging.debug("Exception when calling CustomObjectsApi->create_namespaced_custom_object: %s\n" % e)
+
+    def delete_custom_resource_object(self, application_name):
+        config.load_incluster_config()
+        configuration = client.Configuration()
+        api_instance = client.CustomObjectsApi(client.ApiClient(configuration))
+        group = 'sparkoperator.k8s.io'  # str | The custom resource's group name
+        version = 'v1beta2'  # str | The custom resource's version
+        namespace = 'pnda'  # str | The custom resource's namespace
+        plural = 'sparkapplications'  # str | The custom resource's plural name. For TPRs this would be lowercase plural kind.
+        name = application_name  # str | the custom object's name
+        body = client.V1DeleteOptions()
+        grace_period_seconds = 56
+        orphan_dependents = True
+
+        try:
+            api_response = api_instance.delete_namespaced_custom_object(
+                group,
+                version,
+                namespace,
+                plural,
+                name,
+                body,
+                grace_period_seconds=grace_period_seconds,
+                orphan_dependents=orphan_dependents)
+            logging.info("CRD delete response %s" % api_response)
+        except ApiException as e:
+            logging.debug("Exception when calling CustomObjectsApi->delete_namespaced_custom_object: %s\n" % e)
 
     def create_component(self, staged_component_path, application_name, user_name, component, properties):
         logging.debug("create_component: %s %s %s %s %s", application_name, user_name, json.dumps(component), properties, staged_component_path)
@@ -146,3 +195,17 @@ class SparkStreamingCreator(Common):
         undo_commands.append('rm  %s\n' % service_script_install_path)
         logging.debug("uninstall commands: %s", undo_commands)
         return {'ssh': undo_commands, 'crdjson': jar_path}
+
+    def destroy_component(self, application_name, create_data):
+        logging.debug("destroy_component: %s %s", application_name, json.dumps(create_data))
+        self.delete_custom_resource_object(application_name)
+        self._control_component(create_data['ssh'])
+
+    def start_component(self, application_name, create_data):
+        logging.debug("start_component: %s %s", application_name, json.dumps(create_data))
+        #._control_component(create_data['start_cmds'])
+        json_path = create_data['crdjson']
+        self.create_custom_resource_object("%s/%s.json" % (json_path,application_name))
+
+    def _control_component(self, cmds):
+        deployer_utils.exec_cmds(cmds)
